Move ember_flag_is_active decorator to website/ember_osf_web/decorators.py.
