+++ website/project/views/file.py
@@ -6,7 +6,7 @@ from flask import request
 from website.util import rubeus
 from website.project.decorators import must_be_contributor_or_public, must_not_be_retracted_registration
 from website.project.views.node import _view_project
-from framework.auth.decorators import ember_flag_is_active
+from website.ember_osf_web.decorators import ember_flag_is_active
 
 @must_not_be_retracted_registration
 @must_be_contributor_or_public
